+++ app/overrides/lib/acts_as_taggable_on/tag_override.rb
@@ -12,8 +12,11 @@ ActsAsTaggableOn::Tag.class_eval do
   scope :ordered, -> { order(name: :asc) }
 
   scope :by_site, -> (site) {
-    joins(:taggings).where(taggings: { tenant: site.id })
-                    .select("DISTINCT ON (tags.name) tags.*")
-                    .reorder("tags.name, tags.taggings_count")
+    tenant_site_ids = [site.id]
+    tenant_site_ids << Folio::File.correct_site(site).id if Rails.application.config.folio_shared_files_between_sites
+
+    joins(:taggings).where(taggings: { tenant: tenant_site_ids.uniq })
+                    .select("DISTINCT ON (tags.name) tags.*") # TODO: move to auctocomplete.select2 action?
+                    .reorder("tags.name, tags.taggings_count") # TODO: move to auctocomplete.select2 action?
   }
 end